Move routing module and all pages
Move routing module to sub directory
Move all pages to that directory
